+++ src/main/java/nl/lexemmens/podman/config/image/AbstractImageBuildConfiguration.java
@@ -123,6 +123,15 @@ public abstract class AbstractImageBuildConfiguration {
     protected Boolean squash;
 
     /**
+     * Squash all of the new image’s layers (including those inherited from a base image) into a single new layer.
+     * <p>
+     *
+     * @see "http://docs.podman.io/en/latest/markdown/podman-build.1.html"
+     */
+    @Parameter
+    protected Boolean squashAll;
+
+    /**
      * Cache intermediate images during the build process (Default is true).
      * <p>
      * Note: You can also override the default value of layers by setting the BUILDAH_LAYERS environment variable.
@@ -186,14 +195,6 @@ public abstract class AbstractImageBuildConfiguration {
 
         if (format == null) {
             format = OCI;
-        }
-
-        if (layers == null) {
-            layers = true;
-        }
-
-        if (squash == null) {
-            squash = false;
         }
 
         this.mavenProjectVersion = project.getVersion();
@@ -290,7 +291,16 @@ public abstract class AbstractImageBuildConfiguration {
      *
      * @return true when all of the images's new layers should be squashed into a new layer. False otherwise.
      */
-    public boolean isSquash() {
+    public Boolean getSquash() {
+        return squash;
+    }
+
+    /**
+     * Returns true when all of the image’s new layers, including those inherited from a base image, into a single new layer.
+     *
+     * @return true when all of the images's new layers should be squashed into a new layer. False otherwise.
+     */
+    public Boolean getSquashAll() {
         return squash;
     }
 
@@ -299,7 +309,7 @@ public abstract class AbstractImageBuildConfiguration {
      *
      * @return true, when intermediate layers should be cached. False otherwise.
      */
-    public boolean isLayers() {
+    public Boolean getLayers() {
         return layers;
     }